Added viewing upcoming appointments and cancellation logic

+++ Reserve/Reserve/Endpoints/AppointmentEndpoints.cs
@@ -41,6 +41,11 @@ public static class AppointmentEndpoints
             try
             {
                 await _antiforgery.ValidateRequestAsync(context);
+                Availability deletedSlot = await _appointmentRepository.GetSlotByIdAsync(id);
+                if(deletedSlot.Available == false)
+                {
+                    return Results.BadRequest("Can't delete a reserved slot");
+                }
                 await _appointmentRepository.DeleteAppointmentSlotAsync(id);
                 return Results.Ok();
             }
@@ -60,6 +65,34 @@ public static class AppointmentEndpoints
                     return Results.BadRequest();
                 }
                 return Results.Content($"<tr>\r\n                          <td>{availabilitySlot.StartTime}</td>\r\n                <td>{availabilitySlot.EndTime}</td>\r\n                <td><button hx-delete=\"/delete-slot/{availabilitySlot.Id}\" hx-headers='js:{{\"X-CSRF-TOKEN\": document.getElementsByName(\"__RequestVerificationToken\")[0].value}}' hx-target=\"closest tr\" hx-swap=\"outerHTML swap:1s\" class=\"btn reserve-red-button\">Delete Slot</button></td>\r\n            </tr>", "text/html");
+            }
+            catch (Exception e)
+            {
+                return Results.BadRequest(e.Message);
+            }
+        });
+        group.MapDelete("cancel-appointment/{id}", async (string id, HttpContext context, IAntiforgery _antiforgery, IAppointmentRepository _appointmentRepository) =>
+        {
+            try
+            {
+                await _antiforgery.ValidateRequestAsync(context);
+                AppointmentDetails cancelledAppointment = await _appointmentRepository.GetAppointmentDetailsByIdAsync(id);
+                await _appointmentRepository.CancelAppointmentAsync(cancelledAppointment);
+                context.Response.Headers["HX-Redirect"] = "/Event/CancelNotification";
+                return Results.Ok();
+            }
+            catch (Exception e)
+            {
+                return Results.BadRequest(e.Message);
+            }
+        });
+        group.MapDelete("delete-notification/{id}", async (string id, HttpContext context, IAntiforgery _antiforgery, IAppointmentRepository _appointmentRepository) =>
+        {
+            try
+            {
+                await _antiforgery.ValidateRequestAsync(context);
+                await _appointmentRepository.DeleteNotification(id);
+                return Results.Ok();
             }
             catch (Exception e)
             {
